refactor(core): integrate file serialization into GroupedActionDump
- Add serializeToFiles() instance method to GroupedActionDump
- Add static fromFilesAsInlineJson() method to read and inline screenshots
- Add static cleanupFiles() and getFilePaths() helper methods
- Remove standalone screenshot-io.ts utility (now part of GroupedActionDump)
- Simplify ai-fixture.ts and reporter/index.ts to use new API

This makes file-based serialization a first-class feature of
GroupedActionDump rather than a separate utility, improving cohesion.
