+++ packages/core/src/dump/index.ts
@@ -13,7 +13,3 @@ export {
   generateImageScriptTag,
   generateDumpScriptTag,
 } from './html-utils';
-export {
-  writeScreenshotsToFiles,
-  buildImageMapFromFiles,
-} from './screenshot-io';
