+++ packages/core/src/index.ts
@@ -62,8 +62,6 @@ export {
   parseDumpScriptAttributes,
   generateImageScriptTag,
   generateDumpScriptTag,
-  writeScreenshotsToFiles,
-  buildImageMapFromFiles,
 } from './dump';
 
 // Report generator
